Add systest dev mode and misc review fixes
1. Add contiv-compose add-systest '--start' option to start containers
from go path binary location
2. Make k8s builds build binary from local docker
3. Make k8s-dev starts netplugin from binary
4. Removed no longer required steps from contivNet.sh
5. Not sync bin directory if it's not a test mode

+++ install/k8s/cluster/k8smaster_centos.sh
@@ -5,14 +5,19 @@ if [ -n "$CONTIV_TEST" ]; then
     cp /etc/kubernetes/admin.conf /shared/admin.conf
     chmod 0644 /etc/kubernetes/admin.conf
     cd /opt/gopath/src/github.com/contiv/netplugin/install/k8s/contiv/
-    ./contiv-compose add-systest --k8s-api https://$2:$3 ./contiv-base.yaml > /shared/contiv.yaml
-    # remove kube-dns
-    # TODO: enable kube-dns
-    kubectl delete deployment -n kube-system kube-dns
+
+    if [ "$CONTIV_TEST" = "sys" ]; then
+        ./contiv-compose add-systest --k8s-api https://$2:$3 ./contiv-base.yaml > /shared/contiv.yaml
+        # remove kube-dns
+        # TODO: enable kube-dns
+        kubectl delete deployment -n kube-system kube-dns
+    elif [ "$CONTIV_TEST" = "dev" ]; then
+        ./contiv-compose add-systest --start --k8s-api https://$2:$3 ./contiv-base.yaml > /shared/contiv.yaml
+    fi
 else
     # update to use released version
     cd /opt/gopath/src/github.com/contiv/netplugin/install/k8s/contiv/
-    ./contiv-compose use-release --k8s-api https://$2:$3 -v $(cat /opt/gopath/src/github.com/contiv/netplugin/version/CURRENT_VERSION) ./contiv-base.yaml > /shared/contiv.yaml
+    ./contiv-compose use-release --k8s-api https://$2:$3 -v $(cat ../../../version/CURRENT_VERSION) ./contiv-base.yaml > /shared/contiv.yaml
 fi
 
 kubectl apply -f /shared/contiv.yaml
